requirements.txt
python-dotenv
Jinja2
rich
